pyproject.toml
[tool.poetry]
name = "campusconnect-backend"
version = "0.1.0"
description = "Backend for CampusConnect College Project"
authors = ["Nishant Sapkota <hello@snishant.com.np>"]
license = "MIT"
readme = "README.md"
packages = [{include = "campusconnect_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
pyjwt = "^2.6.0"
django-jazzmin = "^2.6.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
